requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
azure-cognitiveservices-speech==1.41.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect2==0.7.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
lxml==5.3.0
Markdown==3.7
oscrypto==1.3.0
pdfkit==1.0.0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pyparsing==3.2.0
pypdf==5.1.0
python-bidi==0.6.3
python-docx==1.1.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zipp==3.21.0
